77 lines

+++ telegen.py
@@ -10,45 +10,31 @@ import subprocess
 from urlparse import urlparse
 
 class Message:
-
-    procId = os.getpid()
-    p = psutil.Process(procId)
-
     pInfo = type('', (), {})()
-    pInfo.processId = procId
+    pInfo.processId = os.getpid()
+    p = psutil.Process(pInfo.processId)
     pInfo.processName = p.name()
     pInfo.processCmdLine = p.cmdline()
-    pInfo.machine = platform.uname()[1]
     pInfo.username = p.username()
+    pInfo.machine = platform.uname()[1]
     pInfo.currentFolder = p.cwd()
 
-    def __init__(self, domain, action):
-        self.domain = domain
-        self.action = action
-
-    def __str__(self):
-        return jsonpickle.encode(self)
-
 def get_response(msg):
-    resp = jsonpickle.decode(jsonpickle.encode(msg))
+    resp = jsonpickle.decode(jsonpickle.encode(msg)) # Messages should always be readonly, so clone the object. Avoid importing copy library.
     resp.timestamp = datetime.datetime.utcnow().strftime("%Y-%m-%d %H:%M:%S.%f")
     resp.processInfo = Message.pInfo
     return resp
 
 def network_get(msg):
     url = urlparse(msg.url)
-
     sock = socket.socket(socket.AF_INET, socket.SOCK_STREAM)
     sock.settimeout(5)
     sock.connect((url.netloc, url.port or 80))
-
     http = 'GET {0} HTTP/{1} {2}'.format(url.path or '/', 1.0, "\r\n\r\n")
-
     resp = get_response(msg)
     sock.sendall(http)
     resp.bytesSent = len(http)
     sock.close()
-
     resp.protocol = url.scheme
     if hasattr(url, "port"):
         resp.baseurl = url.netloc.split(':')[0]
@@ -59,8 +45,6 @@ def network_get(msg):
     resp.path = url.path
     resp.params = url.params
     resp.query = url.query
-    resp.fragment = url.fragment
-
     return resp
 
 def process_spawn(msg):
@@ -70,23 +54,18 @@ def process_spawn(msg):
 
 def file_update(msg):
     fn = msg.fileName.replace("~", homeFolder)
-    if msg.action == "Delete":
-        os.remove(fn)
+    if msg.action == "Delete": os.remove(fn)
     else:
         f = open(fn, "a+")
-        if hasattr(msg, "append"):
-            f.write(msg.append)
-        if hasattr(msg, "appendline"):
-            f.write(msg.appendline + "\n")
+        if hasattr(msg, "append"): f.write(msg.append)
+        if hasattr(msg, "appendline"): f.write(msg.appendline + "\n")
         f.close()  # just create the file!
     resp = get_response(msg)
     resp.fileName = fn  # update to full path
     return resp
 
 handlers = dict(
-    FileCreate=file_update,
-    FileUpdate=file_update,
-    FileDelete=file_update,
+    FileCreate=file_update, FileUpdate=file_update, FileDelete=file_update,
     NetworkGet=network_get,
     ProcessSpawn=process_spawn
 )
